simplification stack elk + creation of a simple Makefile

+++ configure-system.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
-
 sudo sysctl -w vm.max_map_count=262144
 
-echo "Configurant vm.max_map_count de manière permanente..."
-sudo cp ./ELK/elasticsearch-sysctl.conf /etc/sysctl.d/60-elasticsearch.conf
-sudo sysctl --system
+echo "Configuring vm.max_map_count permanently..."
+if [ ! -f ./ELK/elasticsearch-sysctl.conf ]; then
+    mkdir -p ./ELK
+    echo "vm.max_map_count=262144" > ./ELK/elasticsearch-sysctl.conf    fi
 
-echo "Configuration permanente terminée."
-echo "Vous n'aurez plus besoin d'exécuter sysctl -w vm.max_map_count=262144 après un redémarrage."
+    sudo cp ./ELK/elasticsearch-sysctl.conf /etc/sysctl.d/60-elasticsearch.conf
+    sudo sysctl --system
+
+echo "Permanent configuration complete."
+echo "System is now configured for Elasticsearch."
